goto fonctionnel + prise en compte de la tolérance en dérivée + modif de la condition de stop

+++ src/MotionControlSystem/ControlSettings.h
@@ -37,7 +37,7 @@ public:
     float tolerancyDifferenceSpeed;
 
     uint8_t tolerancyDiferential;
-    uint8_t tolerancyDerivative;
+    float tolerancyDerivative;
 };
 
 #endif //LL_CONTROLSETTINGS_H
